thule-people-service/src/main/java/uk/co/serin/thule/people/CorsRepositoryRestConfigurer.java
package uk.co.serin.thule.people;

import org.springframework.data.rest.core.config.RepositoryRestConfiguration;
import org.springframework.data.rest.webmvc.config.RepositoryRestConfigurer;
import org.springframework.stereotype.Service;

@Service
public class CorsRepositoryRestConfigurer implements RepositoryRestConfigurer {
    @Override
    public void configureRepositoryRestConfiguration(RepositoryRestConfiguration config) {
        config.getCorsRegistry().addMapping("/**").allowedMethods("*").allowedOrigins("*");
    }
}